Folder structure (3rd iteration)
PCOR-xxx

Removed view and datahandler folders from domain.
Updated readme file.
