+++ invoker/domain/userlogin/interface/authviewif.hpp
@@ -18,19 +18,35 @@
 *           Ben Ziv <pointonsoftware@gmail.com>                                                   *
 *                                                                                                 *
 **************************************************************************************************/
-#ifndef INVOKER_DOMAIN_INTERFACE_DATAPROVIDER_DATAPROVIDERIF_HPP_
-#define INVOKER_DOMAIN_INTERFACE_DATAPROVIDER_DATAPROVIDERIF_HPP_
+#ifndef INVOKER_DOMAIN_USERLOGIN_INTERFACE_AUTHVIEWIF_HPP_
+#define INVOKER_DOMAIN_USERLOGIN_INTERFACE_AUTHVIEWIF_HPP_
+#include <string>
 
 namespace domain {
+namespace authentication {
 
-class DataProviderIface {
+class AuthViewIface {
  public:
-    DataProviderIface() = default;
-    virtual ~DataProviderIface() = default;
+    AuthViewIface() = default;
+    virtual ~AuthViewIface() = default;
 
-    virtual bool getDatabaseStatus() = 0;
+    /**
+    * showLoginScreen
+    */
+    virtual void showLoginScreen() = 0;
+
+    /**
+    * showLoginFailed
+    */
+    virtual void showLoginFailed() = 0;
+
+    /**
+    * Will be called if user was found
+    * @param employee a JSON string containing employee details
+    */
+    virtual void loginSuccessful(const std::string& employee) = 0;
 };
 
+}  // namespace authentication
 }  // namespace domain
-
-#endif  // INVOKER_DOMAIN_INTERFACE_DATAPROVIDER_DATAPROVIDERIF_HPP_
+#endif  // INVOKER_DOMAIN_USERLOGIN_INTERFACE_AUTHVIEWIF_HPP_
